Rename processInput
