+++ src/print_help.py
@@ -2,6 +2,6 @@
 #
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
-from pathpicker import processInput
+import process_input
 
-processInput.usage()
+process_input.usage()
